requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --python 3.11 requirements.in -o requirements.txt
asn1crypto==1.5.1
    # via snowflake-connector-python
astunparse==1.6.3
    # via dlt
certifi==2024.2.2
    # via
    #   requests
    #   snowflake-connector-python
cffi==1.16.0
    # via
    #   cryptography
    #   snowflake-connector-python
charset-normalizer==3.3.2
    # via
    #   requests
    #   snowflake-connector-python
click==8.1.7
    # via dlt
cryptography==42.0.7
    # via
    #   pyopenssl
    #   snowflake-connector-python
dlt==0.4.11
duckdb==0.9.2
    # via dlt
filelock==3.14.0
    # via snowflake-connector-python
fsspec==2024.5.0
    # via dlt
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via dlt
giturlparse==0.12.0
    # via dlt
hexbytes==1.2.0
    # via dlt
humanize==4.9.0
    # via dlt
idna==3.7
    # via
    #   requests
    #   snowflake-connector-python
jsonpath-ng==1.6.1
    # via dlt
makefun==1.15.2
    # via dlt
orjson==3.9.10
    # via dlt
packaging==24.0
    # via
    #   dlt
    #   snowflake-connector-python
pathvalidate==3.2.0
    # via dlt
pendulum==3.0.0
    # via dlt
platformdirs==4.2.2
    # via snowflake-connector-python
ply==3.11
    # via jsonpath-ng
pycparser==2.22
    # via cffi
pyjwt==2.8.0
    # via snowflake-connector-python
pyopenssl==24.1.0
    # via snowflake-connector-python
python-dateutil==2.9.0.post0
    # via
    #   pendulum
    #   time-machine
python-dotenv==1.0.1
pytz==2024.1
    # via
    #   dlt
    #   snowflake-connector-python
pyyaml==6.0.1
    # via dlt
requests==2.31.0
    # via
    #   dlt
    #   snowflake-connector-python
requirements-parser==0.9.0
    # via dlt
semver==3.0.2
    # via dlt
setuptools==69.5.1
    # via dlt
simplejson==3.19.2
    # via dlt
six==1.16.0
    # via
    #   astunparse
    #   python-dateutil
smmap==5.0.1
    # via gitdb
snowflake-connector-python==3.10.0
    # via snowflake-sqlalchemy
snowflake-sqlalchemy==1.0.8
sortedcontainers==2.4.0
    # via snowflake-connector-python
sqlalchemy==2.0.30
    # via snowflake-sqlalchemy
tenacity==8.3.0
    # via dlt
time-machine==2.14.1
    # via pendulum
tomlkit==0.12.5
    # via
    #   dlt
    #   snowflake-connector-python
types-setuptools==69.5.0.20240518
    # via requirements-parser
typing-extensions==4.11.0
    # via
    #   dlt
    #   snowflake-connector-python
    #   sqlalchemy
tzdata==2024.1
    # via
    #   dlt
    #   pendulum
urllib3==2.2.1
    # via requests
wheel==0.43.0
    # via astunparse
